refactor(frontend): app layout and event handler (#5279)
