+++ frontend/src/routes/_oh.app/chat-interface.tsx
@@ -1,0 +1,144 @@
+import { useDispatch, useSelector } from "react-redux";
+import React from "react";
+import posthog from "posthog-js";
+import { convertImageToBase64 } from "#/utils/convert-image-to-base-64";
+import { FeedbackActions } from "../../components/feedback-actions";
+import { createChatMessage } from "#/services/chat-service";
+import { InteractiveChatBox } from "../../components/interactive-chat-box";
+import { addUserMessage } from "#/state/chat-slice";
+import { RootState } from "#/store";
+import AgentState from "#/types/agent-state";
+import { generateAgentStateChangeEvent } from "#/services/agent-state-service";
+import { FeedbackModal } from "../../components/feedback-modal";
+import { useScrollToBottom } from "#/hooks/use-scroll-to-bottom";
+import TypingIndicator from "../../components/chat/typing-indicator";
+import { ContinueButton } from "../../components/continue-button";
+import { ScrollToBottomButton } from "../../components/scroll-to-bottom-button";
+import { useWsClient } from "#/context/ws-client-provider";
+import { Messages } from "./messages";
+import { LoadingSpinner } from "./loading-spinner";
+import { ChatSuggestions } from "./chat-suggestions";
+import { ActionSuggestions } from "./action-suggestions";
+
+export function ChatInterface() {
+  const { send, isLoadingMessages } = useWsClient();
+  const dispatch = useDispatch();
+
+  const scrollRef = React.useRef<HTMLDivElement>(null);
+  const { scrollDomToBottom, onChatBodyScroll, hitBottom } =
+    useScrollToBottom(scrollRef);
+
+  const { messages } = useSelector((state: RootState) => state.chat);
+  const { curAgentState } = useSelector((state: RootState) => state.agent);
+
+  const [feedbackPolarity, setFeedbackPolarity] = React.useState<
+    "positive" | "negative"
+  >("positive");
+  const [feedbackModalIsOpen, setFeedbackModalIsOpen] = React.useState(false);
+  const [messageToSend, setMessageToSend] = React.useState<string | null>(null);
+
+  const handleSendMessage = async (content: string, files: File[]) => {
+    posthog.capture("user_message_sent", {
+      current_message_count: messages.length,
+    });
+    const promises = files.map((file) => convertImageToBase64(file));
+    const imageUrls = await Promise.all(promises);
+
+    const timestamp = new Date().toISOString();
+    dispatch(addUserMessage({ content, imageUrls, timestamp }));
+    send(createChatMessage(content, imageUrls, timestamp));
+    setMessageToSend(null);
+  };
+
+  const handleStop = () => {
+    posthog.capture("stop_button_clicked");
+    send(generateAgentStateChangeEvent(AgentState.STOPPED));
+  };
+
+  const handleSendContinueMsg = () => {
+    handleSendMessage("Continue", []);
+  };
+
+  const onClickShareFeedbackActionButton = async (
+    polarity: "positive" | "negative",
+  ) => {
+    setFeedbackModalIsOpen(true);
+    setFeedbackPolarity(polarity);
+  };
+
+  const isWaitingForUserInput =
+    curAgentState === AgentState.AWAITING_USER_INPUT ||
+    curAgentState === AgentState.FINISHED;
+
+  return (
+    <div className="h-full flex flex-col justify-between">
+      {messages.length === 0 && (
+        <ChatSuggestions onSuggestionsClick={setMessageToSend} />
+      )}
+
+      <div
+        ref={scrollRef}
+        onScroll={(e) => onChatBodyScroll(e.currentTarget)}
+        className="flex flex-col grow overflow-y-auto overflow-x-hidden px-4 pt-4 gap-2"
+      >
+        {isLoadingMessages && <LoadingSpinner />}
+
+        {!isLoadingMessages && (
+          <Messages
+            messages={messages}
+            isAwaitingUserConfirmation={
+              curAgentState === AgentState.AWAITING_USER_CONFIRMATION
+            }
+          />
+        )}
+
+        {isWaitingForUserInput && (
+          <ActionSuggestions
+            onSuggestionsClick={(value) => handleSendMessage(value, [])}
+          />
+        )}
+      </div>
+
+      <div className="flex flex-col gap-[6px] px-4 pb-4">
+        <div className="flex justify-between relative">
+          <FeedbackActions
+            onPositiveFeedback={() =>
+              onClickShareFeedbackActionButton("positive")
+            }
+            onNegativeFeedback={() =>
+              onClickShareFeedbackActionButton("negative")
+            }
+          />
+
+          <div className="absolute left-1/2 transform -translate-x-1/2 bottom-0">
+            {messages.length > 2 &&
+              curAgentState === AgentState.AWAITING_USER_INPUT && (
+                <ContinueButton onClick={handleSendContinueMsg} />
+              )}
+            {curAgentState === AgentState.RUNNING && <TypingIndicator />}
+          </div>
+
+          {!hitBottom && <ScrollToBottomButton onClick={scrollDomToBottom} />}
+        </div>
+
+        <InteractiveChatBox
+          onSubmit={handleSendMessage}
+          onStop={handleStop}
+          isDisabled={
+            curAgentState === AgentState.LOADING ||
+            curAgentState === AgentState.AWAITING_USER_CONFIRMATION
+          }
+          mode={curAgentState === AgentState.RUNNING ? "stop" : "submit"}
+          value={messageToSend ?? undefined}
+          onChange={setMessageToSend}
+        />
+      </div>
+
+      <FeedbackModal
+        isOpen={feedbackModalIsOpen}
+        onClose={() => setFeedbackModalIsOpen(false)}
+        polarity={feedbackPolarity}
+      />
+    </div>
+  );
+}
