+++ frontend/src/routes/_oh.app/loading-spinner.tsx
@@ -1,0 +1,7 @@
+export function LoadingSpinner() {
+  return (
+    <div className="flex justify-center">
+      <div className="w-6 h-6 border-2 border-t-[4px] border-primary-500 rounded-full animate-spin" />
+    </div>
+  );
+}
