+++ frontend/src/routes/_oh.app/route.tsx
@@ -2,7 +2,7 @@ import { useDisclosure } from "@nextui-org/react";
 import React from "react";
 import { Outlet } from "@remix-run/react";
 import { useDispatch, useSelector } from "react-redux";
-import Security from "../components/modals/security/security";
+import Security from "#/components/modals/security/security";
 import { Controls } from "#/components/controls";
 import { RootState } from "#/store";
 import { Container } from "#/components/container";
@@ -14,18 +14,20 @@ import GlobeIcon from "#/icons/globe.svg?react";
 import ListIcon from "#/icons/list-type-number.svg?react";
 import { clearJupyter } from "#/state/jupyter-slice";
 import { FilesProvider } from "#/context/files";
-import { ChatInterface } from "#/components/chat-interface";
+import { ChatInterface } from "./chat-interface";
 import { WsClientProvider } from "#/context/ws-client-provider";
-import { EventHandler } from "#/components/event-handler";
+import { EventHandler } from "./event-handler";
 import { useLatestRepoCommit } from "#/hooks/query/use-latest-repo-commit";
 import { useAuth } from "#/context/auth-context";
 import { useUserPrefs } from "#/context/user-prefs-context";
+import { useConversationConfig } from "#/hooks/query/use-conversation-config";
 
 function App() {
   const { token, gitHubToken } = useAuth();
   const { settings } = useUserPrefs();
 
   const dispatch = useDispatch();
+  useConversationConfig();
 
   const { selectedRepository } = useSelector(
     (state: RootState) => state.initalQuery,
@@ -41,7 +43,7 @@ function App() {
   );
 
   const Terminal = React.useMemo(
-    () => React.lazy(() => import("../components/terminal/terminal")),
+    () => React.lazy(() => import("#/components/terminal/terminal")),
     [],
   );
